refactor: 기능 단위로 메서드들을 파일로 분리
